vertical space + move MainAndroid in /Main
