+++ CMakeLists.txt
@@ -19,13 +19,16 @@ lug_set_option(CMAKE_BUILD_TYPE Release STRING "Choose the type of build (Debug 
 if(ANDROID)
     set(ANDROID_ABI arm64-v8a)
     message(STATUS "ANDROID_ABI : ${ANDROID_ABI}")
+
     set(ANDROID_TOOLCHAIN clang)
     message(STATUS "ANDROID_TOOLCHAIN : ${ANDROID_TOOLCHAIN} ")
+
     lug_set_option(ANDROID_PLATFORM android-24 STRING "Target specific android version target (default : android-24)")
     if(NOT ANDROID_PLATFORM MATCHES android-2[3-9])
         message(FATAL_ERROR "Invalid android version target : ${ANDROID_PLATFORM}")
     endif()
     message(STATUS "ANDROID_PLATFORM : ${ANDROID_PLATFORM}")
+
     if(ANDROID_PLATFORM MATCHES android-23)
         message(STATUS "target android-23, forcing static lib")
         set(BUILD_SHARED_LIBS false)
@@ -34,12 +37,15 @@ if(ANDROID)
     if(NOT ANDROID_STL STREQUAL c++_static AND NOT ANDROID_STL STREQUAL c++_shared)
         message(FATAL_ERROR "Invalid STL : ${ANDROID_STL}")
     endif()
+
     if(ANDROID_PLATFORM STREQUAL android-23 AND ANDROID_STL STREQUAL c++_shared)
         message(FATAL_ERROR "android-23 and c++_shared are mutually exclusive")
     endif()
     message(STATUS "ANDROID_STL : ${ANDROID_STL}")
+
     set(ANDROID_CPP_FEATURES "rtti exceptions")
     message(STATUS "ANDROID_CPP_FEATURES : ${ANDROID_CPP_FEATURES}")
+
     set(LIB_SUFFIX "/${ANDROID_ABI}")
 endif()
